Use ClusterName as node-prefix (#76)
* Use ClusterName as node-prefix

This change removes the NodePrefix constants we were using (controlPlaneNode and
workerNode) and instead just uses the ClusterName provided by the spec.  So for
example, now you'll have:
  `<cluster-name>-master-1`
  `<cluster-name>-worker-1`
  `<cluster-name>-worker-2`

This means you can have multiple Rancher servers running on a vSphere (not
overly useful in the real world, but nice for development and testing purposes)

CI testing for example can use a shared vSphere and run multiple tests in
parallel.  A developer can have multiple versions and play around with each one
without tearing down and rebuilding.  It can also be used as a migration tool.

* fix missing string check in rke-cli

* the term master can be offensive, don't use it

* use a single constant form the config pkg

* missed one